requirements.txt
pandas
requests
json
pyyaml
os
